Continued implementation of serialization of the state object.

+++ Cmd/MolTwisterCmdMeasure/CmdAtomPos.cpp
@@ -84,13 +84,13 @@ namespace MolTwisterCmdMeasure
         else if(text1 == "sel")
         {
             fprintf(stdOut_, "\r\n");
-            for(int i=0; i<state_->atoms_.size(); i++)
+            for(int i=0; i<(int)state_->atoms_.size(); i++)
             {
                 if(state_->atoms_[i]->isSelected())
                 {
                     atomPtr = state_->atoms_[i].get();
                     text1 = atomPtr->getID();
-                    if(!(state_->currentFrame_ < atomPtr->r_.size())) continue;
+                    if(!(state_->currentFrame_ < (int)atomPtr->r_.size())) continue;
                     fprintf(stdOut_, "\t[%i,%s,%s]: (%.4f, %.4f, %.4f)\r\n", i, text1.data(), atomPtr->resname_.data(),
                             atomPtr->r_[state_->currentFrame_].x_, atomPtr->r_[state_->currentFrame_].y_, atomPtr->r_[state_->currentFrame_].z_);
                 }
@@ -105,12 +105,12 @@ namespace MolTwisterCmdMeasure
 
         if(foundIndex)
         {
-            if(atomIndex < state_->atoms_.size())
+            if(atomIndex < (int)state_->atoms_.size())
             {
                 atomPtr = state_->atoms_[atomIndex].get();
                 if(atomPtr)
                 {
-                    if(state_->currentFrame_ < atomPtr->r_.size())
+                    if(state_->currentFrame_ < (int)atomPtr->r_.size())
                     {
                         fprintf(stdOut_, "\r\n\t(%.4f, %.4f, %.4f)\r\n", atomPtr->r_[state_->currentFrame_].x_, atomPtr->r_[state_->currentFrame_].y_, atomPtr->r_[state_->currentFrame_].z_);
                     }
